added sliding doors and some new modules

+++ tektor/minecraft/talldoors/doorworkshop/doorparttypes/NullPartType.java
@@ -23,8 +23,8 @@ public class NullPartType extends AbstractDoorPartType{
 	}
 
 	@Override
-	public AbstractDoorPart getNewEntity(World world, int posX,
-			int heightPosition, int posZ, int heightSize, int orientation) {
+	public AbstractDoorPart getNewEntity(World world, double posX,
+			int heightPosition, double posZ, int heightSize, int orientation) {
 		AbstractDoorPart part = new NullPartEntity(world, posX, heightPosition, posZ, heightSize, orientation,depth);
 		return part;
 	}
