+++ tektor/minecraft/talldoors/doorworkshop/doorparttypes/balks/HorizontalBalkPartType.java
@@ -24,8 +24,8 @@ public class HorizontalBalkPartType extends AbstractDoorPartType{
 	}
 
 	@Override
-	public AbstractDoorPart getNewEntity(World world, int posX,
-			int heightPosition, int posZ, int heightSize, int orientation) {
+	public AbstractDoorPart getNewEntity(World world, double posX,
+			int heightPosition, double posZ, int heightSize, int orientation) {
 		AbstractDoorPart part = new HorizontalBalkPartEntity(world, posX, heightPosition, posZ, heightSize, orientation,depth);
 		return part;
 	}
