+++ tektor/minecraft/talldoors/doorworkshop/entity/doorparts/NullPartEntity.java
@@ -9,7 +9,7 @@ public class NullPartEntity extends AbstractDoorPart{
 	public NullPartEntity(World par1World) {
 		super(par1World);
 	}
-	public NullPartEntity(World par1World, int posX, int heightPosition, int posZ,
+	public NullPartEntity(World par1World, double posX, int heightPosition, double posZ,
 			int heightSize, int orientation, float depth) {
 		super(par1World, posX, heightPosition, posZ, heightSize, orientation, depth);
 	}
