+++ tektor/minecraft/talldoors/doorworkshop/entity/doorparts/balks/DoubleHorizontalBalkPartEntity.java
@@ -13,7 +13,7 @@ public class DoubleHorizontalBalkPartEntity extends Abstract2TextureDoorPart{
 
 	
 	
-	public DoubleHorizontalBalkPartEntity(World par1World, int posX, int heightPosition, int posZ,
+	public DoubleHorizontalBalkPartEntity(World par1World, double posX, int heightPosition, double posZ,
 			int heightSize, int orientation, float depth) {
 		super(par1World, posX, heightPosition, posZ, heightSize, orientation, depth);
 	}
