+++ tektor/minecraft/talldoors/doorworkshop/entity/doorparts/balks/HorizontalBalkPartEntity.java
@@ -12,7 +12,7 @@ public class HorizontalBalkPartEntity extends Abstract2TextureDoorPart{
 
 	
 	
-	public HorizontalBalkPartEntity(World par1World, int posX, int heightPosition, int posZ,
+	public HorizontalBalkPartEntity(World par1World, double posX, int heightPosition, double posZ,
 			int heightSize, int orientation, float depth) {
 		super(par1World, posX, heightPosition, posZ, heightSize, orientation, depth);		
 	}
