+++ tektor/minecraft/talldoors/doorworkshop/entity/doorparts/balks/PlusBalkPartEntity.java
@@ -5,8 +5,8 @@ import net.minecraft.world.World;
 
 public class PlusBalkPartEntity extends Abstract2TextureDoorPart{
 
-	public PlusBalkPartEntity(World par1World, int posX, int heightPosition,
-			int posZ, int heightSize, int orientation, float depth) {
+	public PlusBalkPartEntity(World par1World, double posX, int heightPosition,
+			double posZ, int heightSize, int orientation, float depth) {
 		super(par1World, posX, heightPosition, posZ, heightSize, orientation, depth);
 	}
