+++ tektor/minecraft/talldoors/doorworkshop/entity/doorparts/windows/GlassWindowPartEntity.java
@@ -6,8 +6,8 @@ import tektor.minecraft.talldoors.doorworkshop.entity.doorparts.Abstract2Texture
 public class GlassWindowPartEntity extends Abstract2TextureDoorPart {
 
 
-	public GlassWindowPartEntity(World par1World, int posX, int heightPosition,
-			int posZ, int heightSize, int orientation, float depth) {
+	public GlassWindowPartEntity(World par1World, double posX, int heightPosition,
+			double posZ, int heightSize, int orientation, float depth) {
 		super(par1World, posX, heightPosition, posZ, heightSize, orientation,
 				depth);
 	}
