+++ tektor/minecraft/talldoors/doorworkshop/entity/doorparts/windows/HalfedGlassWindowPartEntity.java
@@ -5,8 +5,8 @@ import tektor.minecraft.talldoors.doorworkshop.entity.doorparts.Abstract2Texture
 
 public class HalfedGlassWindowPartEntity extends Abstract2TextureDoorPart {
 
-	public HalfedGlassWindowPartEntity(World par1World, int posX, int heightPosition,
-			int posZ, int heightSize, int orientation, float depth) {
+	public HalfedGlassWindowPartEntity(World par1World, double posX, int heightPosition,
+			double posZ, int heightSize, int orientation, float depth) {
 		super(par1World, posX, heightPosition, posZ, heightSize, orientation, depth);
 	}
