+++ tektor/minecraft/talldoors/doorworkshop/entity/doorparts/windows/SimpleWindowPartEntity.java
@@ -9,7 +9,7 @@ public class SimpleWindowPartEntity extends AbstractDoorPart{
 		super(par1World);
 		}
 
-	public SimpleWindowPartEntity(World par1World, int posX, int heightPosition, int posZ,
+	public SimpleWindowPartEntity(World par1World, double posX, int heightPosition, double posZ,
 			int heightSize, int orientation, float depth) {
 		super(par1World, posX, heightPosition, posZ, heightSize, orientation, depth);
 	}
