+++ tektor/minecraft/talldoors/doorworkshop/entity/doorparts/windows/watchmen/SmallWatchmanWindowEntity.java
@@ -4,8 +4,8 @@ import net.minecraft.world.World;
 
 public class SmallWatchmanWindowEntity extends WatchmanWindowEntity {
 
-	public SmallWatchmanWindowEntity(World par1World, int posX,
-			int heightPosition, int posZ, int heightSize, int orientation,
+	public SmallWatchmanWindowEntity(World par1World, double posX,
+			int heightPosition, double posZ, int heightSize, int orientation,
 			float depth) {
 		super(par1World, posX, heightPosition, posZ, heightSize, orientation, depth);
 	}
